Added error/success messages on change password and delete account page

+++ users/views.py
@@ -115,12 +115,15 @@ def deleteAccountPage(request):
     password1 = request.POST.get('password1')
     password2 = request.POST.get('password2')
 
-    if password1 == password2 and password1 and password2:
-        customer = Customer.objects.get(id=request.user.customer.id)
-        if check_password(password1, customer.user.password):
-            user = User.objects.get(pk=request.user.id)
-            user.delete()
-            return redirect('products:home')
+    if request.POST:
+        if password1 == password2 and password1 and password2:
+            customer = Customer.objects.get(id=request.user.customer.id)
+            if check_password(password1, customer.user.password):
+                user = User.objects.get(pk=request.user.id)
+                user.delete()
+                return redirect('products:home')
+        else:
+            messages.info(request, 'Passwords do not match.')
 
     return render(request, 'users/account/deleteAccount.html')
 
@@ -128,16 +131,18 @@ def deleteAccountPage(request):
 @login_required(login_url='users:login')
 def changePasswordPage(request):
     customer = Customer.objects.get(id=request.user.customer.id)
-    oldPassword = request.POST.get('oldPassword')
     newPassword1 = request.POST.get('newPassword1')
     newPassword2 = request.POST.get('newPassword2')
 
-    if check_password(oldPassword, customer.user.password) and newPassword1 and newPassword2 and newPassword1 == newPassword2:
-        customer.user.set_password(newPassword1)
-        customer.user.save()
-        messages.success(request, 'Password succesfully updated.')
-        request.user.set_password(newPassword1)
-        update_session_auth_hash(request, customer.user)
+    if request.POST:
+        if newPassword1 and newPassword2 and newPassword1 == newPassword2:
+            customer.user.set_password(newPassword1)
+            customer.user.save()
+            messages.success(request, 'Password succesfully updated.')
+            request.user.set_password(newPassword1)
+            update_session_auth_hash(request, customer.user)
+        else:
+            messages.info(request, 'Passwords do not match.')
 
     return render(request, 'users/account/changePassword.html')
